+++ PowerShell/genEclipseRuns.ps1
@@ -6,28 +6,32 @@ if ($MCGradleArg -ne "FromHub")
     # Clear the screen
     Clear-Host
 
+    # Get current PowerShell title (Windows Only)
+    if ($PSVersionTable.Platform -eq "Win32NT")
+    {
+        $MCCurrentTitle = [System.Console]::Title
+    }
+
     $MCGradleAuthor = "Jonathing"
-    $MCGradleVersion = "0.5.0"
+    $MCGradleVersion = "0.5.1"
 
     # Print script information
-    $MCGradleGreeting1 = "MCGradle Scripts"
+    $MCGradleGreeting1 = "MCGradle Scripts by " + $MCGradleAuthor
     $MCGradleGreeting2 = "Version " + $MCGradleVersion
-    $MCGradleGreeting3 = "Written and Maintained by " + $MCGradleAuthor
     Write-Host $MCGradleGreeting1
     Write-Host $MCGradleGreeting2
-    Write-Host $MCGradleGreeting3
     Write-Host ""
 
     # Go to root project directory
     Set-Location ..\..
 
     # Check for update
-    & '.\Scripts\PowerShell Scripts\internal\check_update.ps1' $MCGradleVersion
+    & '.\Scripts\PowerShell\internal\check_update.ps1' $MCGradleVersion
 
     # Get Forge mod name
-    & '.\Scripts\PowerShell Scripts\internal\get_mod_name.ps1'
-    $MCProjectName = Get-Content '.\Scripts\PowerShell Scripts\internal\MODNAME'
-    Remove-Item '.\Scripts\PowerShell Scripts\internal\MODNAME'
+    & '.\Scripts\PowerShell\internal\get_mod_name.ps1'
+    $MCProjectName = Get-Content '.\Scripts\PowerShell\internal\MODNAME'
+    Remove-Item '.\Scripts\PowerShell\internal\MODNAME'
 }
 
 # Set the title of the Windows PowerShell or PowerShell Core console
@@ -39,15 +43,10 @@ $MCTask2Message = "Generating the Eclipse run configurations for " + $MCProjectN
 Write-Host $MCTask2Message
 Write-Host ""
 .\gradlew genEclipseRuns --warning-mode none
+[System.Console]::Title = $MCGradleTitle
 Write-Host ""
 $MCExitMessage = "Finished generating the Eclipse run configurations."
 Write-Host $MCExitMessage
-
-if ($MCGradleArg -ne "FromHub")
-{
-    # Return to scripts directory
-    Set-Location '.\Scripts\PowerShell Scripts\'
-}
 
 # END OF SCRIPT
 Pause
@@ -58,6 +57,33 @@ if ($MCGradleArg -eq "FromHub")
     $MCGradleTitle = $MCProjectName + ": MCGradle Scripts Hub"
     [System.Console]::Title = $MCGradleTitle
 }
+else
+{
+    Write-Host "Quitting MCGradle Scripts..." -ForegroundColor Red
 
-Write-Host ""
+    # Return to scripts directory
+    if ($MCGradleArg -eq "FromCMD")
+    {
+        Set-Location '.\Scripts\Windows\'
+    }
+    else
+    {
+        Set-Location '.\Scripts\PowerShell\'
+    }
+
+    # Revert PowerShell title (Windows Only)
+    if ($PSVersionTable.Platform -eq "Win32NT")
+    {
+        [System.Console]::Title = $MCCurrentTitle
+    }
+    else
+    {
+        [System.Console]::Title = ""
+    }
+}
+
+if ($MCGradleArg -ne "FromCMD")
+{
+    Write-Host ""
+}
 exit 0